fastapi-neon/pyproject.toml
[tool.poetry]
name = "fastapi-neon"
version = "0.1.0"
description = ""
authors = ["DevHussnain <syedhussnainkhalidshah444@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
sqlmodel = "^0.0.16"
psycopg = {extras = ["binary"], version = "^3.1.18"}
pytest = "^8.0.2"
httpx = "^0.27.0"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bycrpt"], version = "^1.7.4"}
types-passlib = "^1.7.7.20240311"
types-python-jose = "^3.3.4.20240106"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
